+++ common/src/main/java/com/lowdragmc/lowdraglib/client/renderer/impl/IModelRenderer.java
@@ -5,19 +5,16 @@ import com.lowdragmc.lowdraglib.client.model.ModelFactory;
 import com.lowdragmc.lowdraglib.client.renderer.IBlockRendererProvider;
 import com.lowdragmc.lowdraglib.client.renderer.IItemRendererProvider;
 import com.lowdragmc.lowdraglib.client.renderer.ISerializableRenderer;
-import com.lowdragmc.lowdraglib.client.renderer.block.RendererBlock;
-import com.lowdragmc.lowdraglib.client.renderer.block.RendererBlockEntity;
 import com.lowdragmc.lowdraglib.gui.editor.ColorPattern;
 import com.lowdragmc.lowdraglib.gui.editor.annotation.ConfigSetter;
 import com.lowdragmc.lowdraglib.gui.editor.annotation.Configurable;
 import com.lowdragmc.lowdraglib.gui.editor.annotation.LDLRegisterClient;
 import com.lowdragmc.lowdraglib.gui.editor.configurator.ConfiguratorGroup;
+import com.lowdragmc.lowdraglib.gui.editor.configurator.StringConfigurator;
 import com.lowdragmc.lowdraglib.gui.editor.configurator.WrapperConfigurator;
 import com.lowdragmc.lowdraglib.gui.editor.ui.Editor;
 import com.lowdragmc.lowdraglib.gui.texture.*;
 import com.lowdragmc.lowdraglib.gui.widget.*;
-import com.lowdragmc.lowdraglib.utils.BlockInfo;
-import com.lowdragmc.lowdraglib.utils.TrackedDummyWorld;
 import com.mojang.blaze3d.vertex.PoseStack;
 import lombok.Getter;
 import net.fabricmc.api.EnvType;
@@ -240,42 +237,37 @@ public class IModelRenderer implements ISerializableRenderer {
         }
     }
 
+    @Environment(EnvType.CLIENT)
+    public void updateModelWithReloadingResource(ResourceLocation modelLocation) {
+        updateModelWithoutReloadingResource(modelLocation);
+        var unBakedModel = getModel();
+        if (unBakedModel == ModelFactory.getUnBakedModel(ModelBakery.MISSING_MODEL_LOCATION)) {
+            Minecraft.getInstance().reloadResourcePacks();
+        }
+    }
+
     @Override
-    public void createPreview(ConfiguratorGroup father) {
-        father.addConfigurators(new WrapperConfigurator("ldlib.gui.editor.group.preview", wrapper -> {
-            var preview = new WidgetGroup(0, 0, 100, 120);
-            var level = new TrackedDummyWorld();
-            level.addBlock(BlockPos.ZERO, BlockInfo.fromBlock(RendererBlock.BLOCK));
-            Optional.ofNullable(level.getBlockEntity(BlockPos.ZERO)).ifPresent(blockEntity -> {
-                if (blockEntity instanceof RendererBlockEntity holder) {
-                    holder.setRenderer(this);
-                }
-            });
-
-            var sceneWidget = new SceneWidget(0, 0, 100, 100, level);
-            sceneWidget.setRenderFacing(false);
-            sceneWidget.setRenderSelect(false);
-            sceneWidget.createScene(level);
-            sceneWidget.getRenderer().setOnLookingAt(null); // better performance
-            sceneWidget.setRenderedCore(Collections.singleton(BlockPos.ZERO), null);
-            sceneWidget.setBackground(new ColorBorderTexture(2, ColorPattern.T_WHITE.color));
-
-            preview.addWidget(new ButtonWidget(5, 110, 90, 10,
-                    new GuiTextureGroup(ColorPattern.T_GRAY.rectTexture().setRadius(5), new TextTexture("ldlib.gui.editor.tips.select_model")), cd -> {
-                if (Editor.INSTANCE == null) return;
-                File path = new File(Editor.INSTANCE.getWorkSpace(), "models");
-                DialogWidget.showFileDialog(Editor.INSTANCE, "ldlib.gui.editor.tips.select_model", path, true,
-                        DialogWidget.suffixFilter(".json"), r -> {
-                            if (r != null && r.isFile()) {
-                                updateModelWithoutReloadingResource(getModelFromFile(path, r));
-                                wrapper.notifyChanges();
-                            }
-                        });
-            }));
-
-            preview.addWidget(sceneWidget);
-            return preview;
-        }));
+    public void buildConfigurator(ConfiguratorGroup father) {
+        ISerializableRenderer.super.buildConfigurator(father);
+        var locationConfigurator = father.getConfigurators().stream()
+                .filter(configurator -> configurator instanceof StringConfigurator stringConfigurator && stringConfigurator.getName().equals("modelLocation"))
+                .map(configurator -> (StringConfigurator) configurator)
+                .findFirst();
+        father.addConfigurators(new WrapperConfigurator(wrapper -> new ButtonWidget(0, 0, 90, 10,
+                new GuiTextureGroup(ColorPattern.T_GRAY.rectTexture().setRadius(5), new TextTexture("ldlib.gui.editor.tips.select_model")), cd -> {
+            if (Editor.INSTANCE == null) return;
+            File path = new File(Editor.INSTANCE.getWorkSpace(), "models");
+            DialogWidget.showFileDialog(Editor.INSTANCE, "ldlib.gui.editor.tips.select_model", path, true,
+                    DialogWidget.suffixFilter(".json"), r -> {
+                        if (r != null && r.isFile()) {
+                            var newModel = getModelFromFile(path, r);
+                            if (newModel.equals(modelLocation)) return;
+                            locationConfigurator.ifPresent(stringConfigurator -> stringConfigurator.setValue(newModel.toString()));
+                            updateModelWithReloadingResource(newModel);
+                            wrapper.notifyChanges();
+                        }
+                    });
+        })).setRemoveTitleBar(true));
     }
 
     private static ResourceLocation getModelFromFile(File path, File r){