+++ common/src/main/java/com/lowdragmc/lowdraglib/gui/editor/ui/MainPanel.java
@@ -13,13 +13,11 @@ import com.lowdragmc.lowdraglib.gui.widget.Widget;
 import com.lowdragmc.lowdraglib.gui.widget.WidgetGroup;
 import com.lowdragmc.lowdraglib.utils.Position;
 import com.lowdragmc.lowdraglib.utils.Size;
-import com.mojang.blaze3d.vertex.PoseStack;
 import lombok.Getter;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.nbt.CompoundTag;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nonnull;
 import java.util.*;
@@ -122,6 +120,9 @@ public class MainPanel extends WidgetGroup {
         for (UIWrapper selectedUI : selectedUIs) {
             selectedUI.remove();
         }
+        if (!selectedUIs.isEmpty()) {
+            editor.addAutoHistory("ldlib.gui.editor.menu.remove", Arrays.hashCode(selectedUIs.toArray()));
+        }
         hoverUI = null;
         selectedUIs.clear();
     }
@@ -151,6 +152,7 @@ public class MainPanel extends WidgetGroup {
             for (UIWrapper selectedUI : selectedUIs) {
                 selectedUI.inner().widget().getParent().onWidgetRemoved(selectedUI.inner());
             }
+            editor.addAutoHistory("ldlib.gui.editor.menu.cut", Arrays.hashCode(selectedUIs.toArray()));
             selectedUIs.clear();
         }
     }
@@ -187,6 +189,7 @@ public class MainPanel extends WidgetGroup {
                         }
                     }
                 }
+                editor.addAutoHistory("ldlib.gui.editor.menu.paste", Arrays.hashCode(list.toArray()));
                 UIResourceTexture.clearCurrentResource();
             });
         }
@@ -201,6 +204,7 @@ public class MainPanel extends WidgetGroup {
             for (UIWrapper ui : selectedUIs) {
                 ui.inner().widget().addSelfPosition(0, max - ui.inner().widget().getRect().down);
             }
+            editor.addAutoHistory("ldlib.gui.editor.menu.align.vb", Arrays.hashCode(selectedUIs.toArray()));
         }
     }
 
@@ -213,6 +217,7 @@ public class MainPanel extends WidgetGroup {
             for (UIWrapper ui : selectedUIs) {
                 ui.inner().widget().addSelfPosition(0, min - ui.inner().widget().getRect().up);
             }
+            editor.addAutoHistory("ldlib.gui.editor.menu.align.vt", Arrays.hashCode(selectedUIs.toArray()));
         }
     }
 
@@ -225,6 +230,7 @@ public class MainPanel extends WidgetGroup {
                 var ui = uis.get(i);
                 ui.addSelfPosition(0, centerY - ui.getRect().getHeightCenter());
             }
+            editor.addAutoHistory("ldlib.gui.editor.menu.align.vd", Arrays.hashCode(selectedUIs.toArray()));
         }
     }
 
@@ -239,6 +245,7 @@ public class MainPanel extends WidgetGroup {
             for (UIWrapper ui : selectedUIs) {
                 ui.inner().widget().addSelfPosition(0, mid - ui.inner().widget().getRect().getHeightCenter());
             }
+            editor.addAutoHistory("ldlib.gui.editor.menu.align.vc", Arrays.hashCode(selectedUIs.toArray()));
         }
     }
 
@@ -251,6 +258,7 @@ public class MainPanel extends WidgetGroup {
             for (UIWrapper ui : selectedUIs) {
                 ui.inner().widget().addSelfPosition(max - ui.inner().widget().getRect().right, 0);
             }
+            editor.addAutoHistory("ldlib.gui.editor.menu.align.hr", Arrays.hashCode(selectedUIs.toArray()));
         }
     }
 
@@ -263,6 +271,7 @@ public class MainPanel extends WidgetGroup {
             for (UIWrapper ui : selectedUIs) {
                 ui.inner().widget().addSelfPosition(min - ui.inner().widget().getRect().left, 0);
             }
+            editor.addAutoHistory("ldlib.gui.editor.menu.align.hl", Arrays.hashCode(selectedUIs.toArray()));
         }
     }
 
@@ -275,6 +284,7 @@ public class MainPanel extends WidgetGroup {
                 var ui = uis.get(i);
                 ui.addSelfPosition(centerX - ui.getRect().getWidthCenter(), 0);
             }
+            editor.addAutoHistory("ldlib.gui.editor.menu.align.hd", Arrays.hashCode(selectedUIs.toArray()));
         }
     }
 
@@ -289,6 +299,7 @@ public class MainPanel extends WidgetGroup {
             for (UIWrapper ui : selectedUIs) {
                 ui.inner().widget().addSelfPosition(mid - ui.inner().widget().getRect().getWidthCenter(), 0);
             }
+            editor.addAutoHistory("ldlib.gui.editor.menu.align.hc", Arrays.hashCode(selectedUIs.toArray()));
         }
     }
 
@@ -348,11 +359,12 @@ public class MainPanel extends WidgetGroup {
                 for (UIWrapper selectedUI : selectedUIs) {
                     selectedUI.onDragPosition((int) deltaX, (int) deltaY);
                 }
+                editor.addAutoHistory("ldlib.gui.editor.dragging_position", Arrays.hashCode(selectedUIs.toArray()));
             } else if (isDragSize) {
                 for (UIWrapper selectedUI : selectedUIs) {
                     selectedUI.onDragSize((int) deltaX, (int) deltaY);
-
                 }
+                editor.addAutoHistory("ldlib.gui.editor.dragging_size", Arrays.hashCode(selectedUIs.toArray()));
             }
             return true;
         }