+++ common/src/main/java/com/lowdragmc/lowdraglib/test/ui/TestSceneObject.java
@@ -2,12 +2,13 @@ package com.lowdragmc.lowdraglib.test.ui;
 
 import com.lowdragmc.lowdraglib.gui.editor.annotation.LDLRegisterClient;
 import com.lowdragmc.lowdraglib.gui.editor.ui.sceneeditor.SceneEditorWidget;
-import com.lowdragmc.lowdraglib.gui.editor.ui.sceneeditor.sceneobject.utils.SizeBoxObject;
+import com.lowdragmc.lowdraglib.gui.editor.ui.sceneeditor.sceneobject.utils.TransformGizmoObject;
 import com.lowdragmc.lowdraglib.gui.modular.IUIHolder;
 import com.lowdragmc.lowdraglib.gui.modular.ModularUI;
 import lombok.NoArgsConstructor;
 import net.minecraft.client.Minecraft;
 import net.minecraft.world.entity.player.Player;
+import org.joml.Quaternionf;
 import org.joml.Vector3f;
 
 import java.util.List;
@@ -34,13 +35,14 @@ public class TestSceneObject implements IUITest {
                 playerPos.south()
         ));
 
-//        var anchor = new TransformAnchorObject();
-//        anchor.transform().position(new Vector3f(playerPos.above().getX(), playerPos.above().getY(), playerPos.above().getZ()));
-//        sceneWidget.addSceneObject(anchor);
+        var anchor = new TransformGizmoObject();
+        anchor.transform().position(new Vector3f(playerPos.above().getX(), playerPos.above().getY(), playerPos.above().getZ()));
+        anchor.transform().rotation(new Quaternionf().identity().rotationXYZ(30, 23, 75));
+        sceneWidget.addSceneObject(anchor);
 
-        var sizeBox = new SizeBoxObject();
-        sizeBox.transform().position(new Vector3f(playerPos.below().getX(), playerPos.below().getY() + 1, playerPos.below().getZ()));
-        sceneWidget.addSceneObject(sizeBox);
+//        var sizeBox = new SizeBoxObject();
+//        sizeBox.transform().position(new Vector3f(playerPos.below().getX(), playerPos.below().getY() + 1, playerPos.below().getZ()));
+//        sceneWidget.addSceneObject(sizeBox);
 
         return IUITest.super.createUI(holder, entityPlayer)
                 .widget(sceneWidget);