+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # ChangeLog
 
-* fix loop node crash
+* fix ListMergeNode wont convert compatible types
+* add transform anchor feature
+* some QoL improvements